Reorganize FE Admin components
