+++ frontend/src/components/Admin/EditTeam/EditTeam.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import { Card, Button, Form, Input } from 'semantic-ui-react';
 import styles from './EditTeam.module.css';
-import ErrorModal from '../Modals/ErrorModal';
-import { Emitters } from '../../utils';
-import { Team, TeamsAPI } from '../../API/TeamsAPI';
-import { MemberSearch } from '../Common/Search';
-import { useTeams } from '../Common/FirestoreDataProvider';
+import ErrorModal from '../../Modals/ErrorModal';
+import { Emitters } from '../../../utils';
+import { Team, TeamsAPI } from '../../../API/TeamsAPI';
+import { MemberSearch } from '../../Common/Search';
+import { useTeams } from '../../Common/FirestoreDataProvider';
 
 type TeamMemberEditorProps = {
   readonly title: string;
