+++ frontend/src/pages/admin/candidate-decider.tsx
@@ -1,3 +1,3 @@
-import AdminCandidateDeciderBase from '../../components/Admin/AdminCandidateDecider';
+import AdminCandidateDeciderBase from "../../components/Admin/AdminCandidateDecider/AdminCandidateDecider";
 
 export default AdminCandidateDeciderBase;
