+++ frontend/src/pages/admin/edit.tsx
@@ -1,3 +1,3 @@
-import AddUser from '../../components/Admin/AddUser';
+import AddUser from '../../components/Admin/AddUser/AddUser';
 
 export default AddUser;
