+++ frontend/src/pages/admin/signin-creator.tsx
@@ -1,3 +1,3 @@
-import SignInFormCreatorBase from '../../components/Admin/SignInFormCreator';
+import SignInFormCreatorBase from '../../components/Admin/SignInForm/SignInFormCreator';
 
 export default SignInFormCreatorBase;
