+++ frontend/src/pages/admin/site-deployer.tsx
@@ -1,3 +1,3 @@
-import SiteDeployer from '../../components/Admin/SiteDeployer';
+import SiteDeployer from '../../components/Admin/SiteDeployer/SiteDeployer';
 
 export default SiteDeployer;
