+++ frontend/src/pages/admin/team-event-details/[uuid].tsx
@@ -1,3 +1,3 @@
-import TeamEventDetails from '../../../components/Admin/TeamEventDetails';
+import TeamEventDetails from '../../../components/Admin/TeamEvent/TeamEventDetails';
 
 export default TeamEventDetails;
