+++ frontend/src/pages/admin/team-events.tsx
@@ -1,3 +1,3 @@
-import TeamEvents from '../../components/Admin/TeamEvents';
+import TeamEvents from '../../components/Admin/TeamEvent/TeamEvents';
 
 export default TeamEvents;
